refactor the code to make reviews a seprate component
